fix TJA newline after last note symbol and before `,` caused wrong note-symbols-per-line calculation
The parser assumed that the measure-final `,` must come right before
the last note symbol of a measure,
and just replaced `,` lines with `0,`.

This is not true as TaikoJiro allows newlines between each note symbol and `,`,
and a newline is required if a command is placed right before the `,`.

The original handling was also ignoring every indented `,` line
while currently running completely fine, so the replacing is now unnecessary.

This commit simply stops replacing `,` lines with `0,`.